WIP: fix control flow
