+++ crates/edr_solidity/src/solidity_stack_trace.rs
@@ -1,8 +1,18 @@
 //! Stack trace entries for Solidity errors.
 
-use edr_primitives::{Address, Bytes, U256};
+use std::collections::HashMap;
 
-use crate::{build_model::ContractFunctionType, return_data::CheatcodeErrorDetails};
+use edr_chain_spec::HaltReasonTrait;
+use edr_primitives::{Address, Bytes, U256};
+use revm_inspectors::tracing::CallTraceArena;
+
+use crate::{
+    build_model::ContractFunctionType,
+    contract_decoder::{ContractDecoderError, NestedTraceDecoder},
+    nested_trace::{CallTraceArenaConversionError, NestedTrace},
+    return_data::CheatcodeErrorDetails,
+    solidity_tracer::{self, SolidityTracerError},
+};
 
 pub(crate) const FALLBACK_FUNCTION_NAME: &str = "<fallback>";
 pub(crate) const RECEIVE_FUNCTION_NAME: &str = "<receive>";
@@ -210,6 +220,133 @@ impl StackTraceEntry {
                 address == contract_address
             }
             _ => false,
+        }
+    }
+}
+
+/// Stack trace creation error.
+#[derive(Clone, Debug, thiserror::Error)]
+pub enum StackTraceCreationError<HaltReasonT> {
+    /// Error during contract decoding.
+    #[error(transparent)]
+    ContractDecoder(#[from] ContractDecoderError),
+    /// Error during trace conversion
+    #[error(transparent)]
+    TraceConversion(#[from] CallTraceArenaConversionError),
+    /// Error with the provided input trace.
+    #[error(transparent)]
+    Tracer(#[from] SolidityTracerError<HaltReasonT>),
+}
+
+impl<HaltReasonT> StackTraceCreationError<HaltReasonT> {
+    /// Maps the type of the halt reason using the provided conversion function.
+    pub fn map_halt_reason<
+        ConversionFnT: Copy + Fn(HaltReasonT) -> NewHaltReasonT,
+        NewHaltReasonT,
+    >(
+        self,
+        conversion_fn: ConversionFnT,
+    ) -> StackTraceCreationError<NewHaltReasonT> {
+        match self {
+            StackTraceCreationError::ContractDecoder(err) => {
+                StackTraceCreationError::ContractDecoder(err)
+            }
+            StackTraceCreationError::TraceConversion(err) => {
+                StackTraceCreationError::TraceConversion(err)
+            }
+            StackTraceCreationError::Tracer(err) => {
+                StackTraceCreationError::Tracer(err.map_halt_reason(conversion_fn))
+            }
+        }
+    }
+}
+
+/// Compute stack trace based on execution traces.
+/// Assumes last trace is the error one. This is important for invariant tests
+/// where there might be multiple errors traces. Returns `None` if `traces` is
+/// empty.
+pub fn get_stack_trace<
+    'arena,
+    HaltReasonT: HaltReasonTrait,
+    NestedTraceDecoderT: NestedTraceDecoder<HaltReasonT>,
+>(
+    contract_decoder: &NestedTraceDecoderT,
+    traces: impl IntoIterator<Item = &'arena CallTraceArena>,
+) -> Result<Option<Vec<StackTraceEntry>>, StackTraceCreationError<HaltReasonT>> {
+    let mut address_to_creation_code = HashMap::new();
+    let mut address_to_runtime_code = HashMap::new();
+
+    let last_trace = traces.into_iter().fold(None, |_, trace| {
+        for node in trace.nodes() {
+            let address = node.trace.address;
+            if node.trace.kind.is_any_create() {
+                address_to_creation_code.insert(address, &node.trace.data);
+                address_to_runtime_code.insert(address, &node.trace.output);
+            }
+        }
+        Some(trace)
+    });
+
+    if let Some(last_trace) = last_trace {
+        let trace = NestedTrace::from_call_trace_arena(
+            &address_to_creation_code,
+            &address_to_runtime_code,
+            last_trace,
+        )?;
+        let trace = contract_decoder.try_to_decode_nested_trace(trace)?;
+        let stack_trace = solidity_tracer::get_stack_trace(trace)?;
+        Ok(Some(stack_trace))
+    } else {
+        Ok(None)
+    }
+}
+
+/// The possible outcomes from computing stack traces.
+#[derive(Clone, Debug)]
+pub enum StackTraceCreationResult<HaltReasonT> {
+    /// The stack trace result
+    Success(Vec<StackTraceEntry>),
+    /// We couldn't generate stack traces, because an unexpected error occurred.
+    Error(StackTraceCreationError<HaltReasonT>),
+    /// We couldn't generate stack traces, because the heuristic failed.
+    HeuristicFailed,
+}
+
+impl<HaltReasonT> StackTraceCreationResult<HaltReasonT> {
+    /// Maps the type of the halt reason using the provided conversion function.
+    pub fn map_halt_reason<
+        ConversionFnT: Copy + Fn(HaltReasonT) -> NewHaltReasonT,
+        NewHaltReasonT,
+    >(
+        self,
+        conversion_fn: ConversionFnT,
+    ) -> StackTraceCreationResult<NewHaltReasonT> {
+        match self {
+            StackTraceCreationResult::Success(stack_trace) => {
+                StackTraceCreationResult::Success(stack_trace)
+            }
+            StackTraceCreationResult::Error(error) => {
+                StackTraceCreationResult::Error(error.map_halt_reason(conversion_fn))
+            }
+            StackTraceCreationResult::HeuristicFailed => StackTraceCreationResult::HeuristicFailed,
+        }
+    }
+}
+
+impl<HaltReasonT: HaltReasonTrait>
+    From<Result<Vec<StackTraceEntry>, StackTraceCreationError<HaltReasonT>>>
+    for StackTraceCreationResult<HaltReasonT>
+{
+    fn from(value: Result<Vec<StackTraceEntry>, StackTraceCreationError<HaltReasonT>>) -> Self {
+        match value {
+            Ok(stack_trace) => {
+                if stack_trace.is_empty() {
+                    Self::HeuristicFailed
+                } else {
+                    Self::Success(stack_trace)
+                }
+            }
+            Err(error) => Self::Error(error),
         }
     }
 }
