+++ crates/foundry/evm/evm/src/executors/stack_trace.rs
@@ -1,14 +1,6 @@
-use std::{
-    borrow::Cow,
-    collections::{HashMap, HashSet},
-};
+use std::{borrow::Cow, collections::HashSet};
 
-use edr_solidity::{
-    contract_decoder::{ContractDecoderError, NestedTraceDecoder},
-    nested_trace::{CallTraceArenaConversionError, NestedTrace},
-    solidity_stack_trace::StackTraceEntry,
-    solidity_tracer::{self, SolidityTracerError},
-};
+use edr_solidity::solidity_stack_trace::{StackTraceCreationError, StackTraceEntry};
 use foundry_evm_core::{
     backend::IndeterminismReasons,
     evm_context::{
@@ -16,43 +8,9 @@ use foundry_evm_core::{
         TransactionErrorTrait,
     },
 };
-use foundry_evm_traces::CallTraceArena;
 use revm::context::result::HaltReasonTr;
 
 use crate::executors::{EvmError, ExecutorBuilderError};
-
-/// Stack trace creation error.
-#[derive(Clone, Debug, thiserror::Error)]
-pub enum StackTraceCreationError<HaltReasonT> {
-    #[error(transparent)]
-    ContractDecoder(#[from] ContractDecoderError),
-    #[error(transparent)]
-    TraceConversion(#[from] CallTraceArenaConversionError),
-    #[error(transparent)]
-    Tracer(#[from] SolidityTracerError<HaltReasonT>),
-}
-
-impl<HaltReasonT> StackTraceCreationError<HaltReasonT> {
-    pub fn map_halt_reason<
-        ConversionFnT: Copy + Fn(HaltReasonT) -> NewHaltReasonT,
-        NewHaltReasonT,
-    >(
-        self,
-        conversion_fn: ConversionFnT,
-    ) -> StackTraceCreationError<NewHaltReasonT> {
-        match self {
-            StackTraceCreationError::ContractDecoder(err) => {
-                StackTraceCreationError::ContractDecoder(err)
-            }
-            StackTraceCreationError::TraceConversion(err) => {
-                StackTraceCreationError::TraceConversion(err)
-            }
-            StackTraceCreationError::Tracer(err) => {
-                StackTraceCreationError::Tracer(err.map_halt_reason(conversion_fn))
-            }
-        }
-    }
-}
 
 /// Stack trace generation error during re-execution.
 #[derive(Clone, Debug, thiserror::Error)]
@@ -127,94 +85,6 @@ impl<
     }
 }
 
-/// Compute stack trace based on execution traces.
-/// Assumes last trace is the error one. This is important for invariant tests
-/// where there might be multiple errors traces. Returns `None` if `traces` is
-/// empty.
-pub fn get_stack_trace<
-    'arena,
-    HaltReasonT: HaltReasonTr,
-    NestedTraceDecoderT: NestedTraceDecoder<HaltReasonT>,
->(
-    contract_decoder: &NestedTraceDecoderT,
-    traces: impl IntoIterator<Item = &'arena CallTraceArena>,
-) -> Result<Option<Vec<StackTraceEntry>>, StackTraceCreationError<HaltReasonT>> {
-    let mut address_to_creation_code = HashMap::new();
-    let mut address_to_runtime_code = HashMap::new();
-
-    let last_trace = traces.into_iter().fold(None, |_, trace| {
-        for node in trace.nodes() {
-            let address = node.trace.address;
-            if node.trace.kind.is_any_create() {
-                address_to_creation_code.insert(address, &node.trace.data);
-                address_to_runtime_code.insert(address, &node.trace.output);
-            }
-        }
-        Some(trace)
-    });
-
-    if let Some(last_trace) = last_trace {
-        let trace = NestedTrace::from_call_trace_arena(
-            &address_to_creation_code,
-            &address_to_runtime_code,
-            last_trace,
-        )?;
-        let trace = contract_decoder.try_to_decode_nested_trace(trace)?;
-        let stack_trace = solidity_tracer::get_stack_trace(trace)?;
-        Ok(Some(stack_trace))
-    } else {
-        Ok(None)
-    }
-}
-
-/// The possible outcomes from computing stack traces.
-#[derive(Clone, Debug)]
-pub enum StackTraceCreationResult<HaltReasonT> {
-    /// The stack trace result
-    Success(Vec<StackTraceEntry>),
-    /// We couldn't generate stack traces, because an unexpected error occurred.
-    Error(StackTraceCreationError<HaltReasonT>),
-    HeuristicFailed,
-}
-
-impl<HaltReasonT> StackTraceCreationResult<HaltReasonT> {
-    pub fn map_halt_reason<
-        ConversionFnT: Copy + Fn(HaltReasonT) -> NewHaltReasonT,
-        NewHaltReasonT,
-    >(
-        self,
-        conversion_fn: ConversionFnT,
-    ) -> StackTraceCreationResult<NewHaltReasonT> {
-        match self {
-            StackTraceCreationResult::Success(stack_trace) => {
-                StackTraceCreationResult::Success(stack_trace)
-            }
-            StackTraceCreationResult::Error(error) => {
-                StackTraceCreationResult::Error(error.map_halt_reason(conversion_fn))
-            }
-            StackTraceCreationResult::HeuristicFailed => StackTraceCreationResult::HeuristicFailed,
-        }
-    }
-}
-
-impl<HaltReasonT: HaltReasonTr>
-    From<Result<Vec<StackTraceEntry>, StackTraceCreationError<HaltReasonT>>>
-    for StackTraceCreationResult<HaltReasonT>
-{
-    fn from(value: Result<Vec<StackTraceEntry>, StackTraceCreationError<HaltReasonT>>) -> Self {
-        match value {
-            Ok(stack_trace) => {
-                if stack_trace.is_empty() {
-                    Self::HeuristicFailed
-                } else {
-                    Self::Success(stack_trace)
-                }
-            }
-            Err(error) => Self::Error(error),
-        }
-    }
-}
-
 /// The possible outcomes from trying to compute a stack trace for Solidity
 /// tests.
 #[derive(Clone, Debug)]
@@ -277,17 +147,15 @@ impl<HaltReasonT: HaltReasonTr>
 {
     fn from(value: Result<Vec<StackTraceEntry>, SolidityTestStackTraceError<HaltReasonT>>) -> Self {
         match value {
-            Ok(stack_trace) => Self::Creation(StackTraceCreationResult::with_entries(stack_trace)),
-            Err(error) => Self::Error(error.into()),
+            Ok(stack_trace) => {
+                if stack_trace.is_empty() {
+                    Self::HeuristicFailed
+                } else {
+                    Self::Success(stack_trace)
+                }
+            }
+            Err(error) => Self::Error(error),
         }
-    }
-}
-
-impl<HaltReasonT: HaltReasonTr> From<StackTraceCreationResult<HaltReasonT>>
-    for SolidityTestStackTraceResult<HaltReasonT>
-{
-    fn from(value: StackTraceCreationResult<HaltReasonT>) -> Self {
-        Self::Creation(value)
     }
 }
